pyproject.toml
[project]
name = "vibe-check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "jsonschema>=4.26.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pyarrow>=23.0.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.54.0",
]
